requirements.txt
colorama==0.4.4
cycler==0.11.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.4
Pillow==8.4.0
pyparsing==3.0.5
python-dateutil==2.8.2
six==1.16.0
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.62.3
typing-extensions==3.10.0.2
